refactor: Datasource, Snapshot, API Request ViewModel로 이전
