+++ OpenMarket/OpenMarket/Assets.xcassets/AccentColor.colorset/Contents.json
