+++ OpenMarket/OpenMarket/Assets.xcassets/AppIcon.appiconset/Contents.json
