+++ OpenMarket/OpenMarket/Assets.xcassets/Contents.json
